Use parse_profile for profile setter methods
