+++ desc/equilibrium/utils.py
@@ -1,0 +1,53 @@
+"""Utility functions for parsing inputs to Equilibrium."""
+
+import numbers
+
+import numpy as np
+
+from desc.backend import jnp
+from desc.profiles import PowerSeriesProfile, Profile
+
+
+def parse_profile(prof, name="", **kwargs):
+    """Parse an object into a Profile.
+
+    Parameters
+    ----------
+    prof : numeric, ndarray, Profile
+        Object to parse. Profile objects are returned unaltered. Scalars are treated
+        as a constant profile. Arrays of shape(k,) are treated as coefficients of
+        power series, arrays of shape (k,2) are treated as mode numbers and coefficients
+        of a power series. If None, returns None
+    name : str, optional
+        Name to assign to profile if prof is numeric or ndarray
+    kwargs: any
+        Additional keyword arguments are passed through to Profile constructor if prof
+        is numeric or ndarray.
+
+    Returns
+    -------
+    prof : Profile or None
+        Profile object, or None if input is None
+
+    Raises
+    ------
+    TypeError
+        If the object cannot be parsed as a Profile
+    """
+    if isinstance(prof, Profile):
+        return prof
+    if isinstance(prof, numbers.Number) or (
+        isinstance(prof, (np.ndarray, jnp.ndarray)) and prof.ndim == 1
+    ):
+        return PowerSeriesProfile(params=prof, name=name, **kwargs)
+    if (
+        isinstance(prof, (np.ndarray, jnp.ndarray))
+        and prof.ndim == 2
+        and prof.shape[1] == 2
+    ):
+        return PowerSeriesProfile(
+            modes=prof[:, 0], params=prof[:, 1], name=name, **kwargs
+        )
+    if prof is None:
+        return None
+    raise TypeError(f"Got unknown {name} profile {prof}")
